requirements.txt
Django==1.7
python-social-auth==0.2.7

